backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0

openpyxl>=3.1.0  # Excel template merge (export-to-excel)
openai>=1.0.0
tiktoken>=0.7.0
google-genai>=1.0.0
groq>=1.0.0
json-repair>=0.50.0
